# api/openapi.yaml
openapi: 3.1.0
info:
  title: Island Media Co connected assessment API
  version: 2.0.0
  description: >-
    Fictional mock boundary for the Frameworks technical assessment. Protected
    prototype endpoints use X-Prototype-User-Id to represent the selected
    seeded or newly registered user. Production authentication is outside scope.
servers:
  - url: /api
paths:
...
  /management/contracts:
    post:
      summary: Create a draft contract from an approved request
      parameters:
        - $ref: "#/components/parameters/PrototypeUserId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContractCreate"
      responses:
        "201":
          description: Draft contract created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Contract"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /management/work-orders:
    post:
      summary: Create and assign a campaign work order
      parameters:
        - $ref: "#/components/parameters/PrototypeUserId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderCreate"
      responses:
        "201":
          description: Work order created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
...
  /mobile/work-orders/{workOrderId}/status:
    post:
      summary: Update fitter progress or report a blocker
      parameters:
        - $ref: "#/components/parameters/PrototypeUserId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - in: path
          name: workOrderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkOrderStatusUpdate"
      responses:
        "200":
          description: Updated work order and history
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkOrder"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
        "422":
          $ref: "#/components/responses/ValidationError"
  /mobile/work-orders/{workOrderId}/proof:
    post:
      summary: Attach prototype completion proof
      parameters:
        - $ref: "#/components/parameters/PrototypeUserId"
        - $ref: "#/components/parameters/IdempotencyKey"
        - in: path
          name: workOrderId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, completionNote]
              properties:
                file:
                  type: string
                  format: binary
                completionNote:
                  type: string
                  minLength: 3
      responses:
        "201":
          description: Proof record created and service history updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProofRecord"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
components:
  parameters:
    PrototypeUserId:
      in: header
      name: X-Prototype-User-Id
      required: true
      schema:
        type: string
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      schema:
        type: string
        minLength: 8
  schemas:
...
    HistoryEntry:
      type: object
      required: [at, actor, action]
      properties:
        at:
          type: string
          format: date-time
        actor:
          type: string
        action:
          type: string
        note:
          type: [string, "null"]
...
    ContractItem:
      type: object
      required: [id, productId, quantity, lineTotal]
      properties:
        id:
          type: string
        productId:
          type: string
        assetId:
          type: [string, "null"]
        capacityPoolId:
          type: [string, "null"]
        quantity:
          type: integer
        unitRate:
          type: [number, "null"]
        rateUnit:
          type: [string, "null"]
        lineTotal:
          type: number
    ContractCreate:
      type: object
      required: [organisationId, bookingRequestId, startDate, endDate, items, total]
      properties:
        organisationId:
          type: string
        bookingRequestId:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        items:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/ContractItem"
        total:
          type: number
          minimum: 0
    Contract:
      type: object
      required:
        [
          id,
          organisationId,
          status,
          version,
          startDate,
          endDate,
          currency,
          total,
          items,
          history,
        ]
      properties:
        id:
          type: string
        organisationId:
          type: string
        bookingRequestId:
          type: [string, "null"]
        status:
          type: string
          enum:
            [draft, issued, change_requested, accepted, active, completed, cancelled]
        version:
          type: integer
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        currency:
          type: string
          const: GBP
        total:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/ContractItem"
        history:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
...
    ProofRecord:
      type: object
      required: [id, workOrderId, fileName, createdAt, createdByUserId]
      properties:
        id:
          type: string
        workOrderId:
          type: string
        fileName:
          type: string
        previewUrl:
          type: [string, "null"]
        completionNote:
          type: string
        createdAt:
          type: string
          format: date-time
        createdByUserId:
          type: string
...
    WorkOrderCreate:
      type: object
      required:
        [
          campaignId,
          contractId,
          type,
          assignedUserId,
          assetId,
          scheduledStart,
          scheduledEnd,
          locationLabel,
          instructions,
        ]
      properties:
        campaignId:
          type: string
        contractId:
          type: string
        type:
          type: string
          enum: [survey, production, installation, maintenance, removal]
        assignedUserId:
          type: string
        assetId:
          type: string
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        locationLabel:
          type: string
        instructions:
          type: string
    WorkOrder:
      type: object
      required:
        [
          id,
          campaignId,
          contractId,
          organisationId,
          type,
          status,
          assignedUserId,
          assetId,
          scheduledStart,
          scheduledEnd,
          locationLabel,
          instructions,
          history,
          proofRecordIds,
        ]
      properties:
        id:
          type: string
        campaignId:
          type: string
        contractId:
          type: string
        organisationId:
          type: string
        type:
          type: string
        status:
          type: string
          enum: [draft, assigned, travelling, on_site, blocked, completed]
        assignedUserId:
          type: string
        assetId:
          type: string
        scheduledStart:
          type: string
          format: date-time
        scheduledEnd:
          type: string
          format: date-time
        locationLabel:
          type: string
        instructions:
          type: string
        completionNote:
          type: [string, "null"]
        proofRecordIds:
          type: array
          items:
            type: string
        history:
          type: array
          items:
            $ref: "#/components/schemas/HistoryEntry"
    WorkOrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [assigned, travelling, on_site, blocked, completed]
        note:
          type: [string, "null"]
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: [object, "null"]
          additionalProperties: true
  responses:
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Forbidden:
      description: The selected prototype user cannot access this record or action
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: FORBIDDEN
            message: You cannot access this organisation's record.
    Conflict:
      description: State or availability changed before the action completed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: INVENTORY_CONFLICT
            message: No suitable inventory remains for the requested dates.
    ServiceUnavailable:
      description: Simulated or real temporary service failure
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: SERVICE_UNAVAILABLE
            message: Please try again. The service is temporarily unavailable.